Fix module defaults
Previously I was trying to pass in the CODEBUILD_SRC vars in env
vars. We could do that with something nasty like an eval but
instead I've just passed the path within the artifact and the
artifact name as a separate var.

The bash then figures out the env var name and resolves the path
from it.

The defaults should now work if you don't specify any of these
vars but it means that you can choose to save the files to
artifact names or file names of your choosing.